user-microservice/pyproject.toml
[project]
name = "user-microservice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.121.1",
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
    "sqlalchemy>=2.0",
    "asyncpg>=0.27.0",
    "ruff>=0.14.5",
    "slowapi>=0.1.9",
    "alembic>=1.17.2",
    "psycopg2-binary>=2.9.11",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "redis>=5.2.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
